Add files via upload
Hi everybody! I made some modifications in the code that includes some functions to allow us to use more than one method, for example. Our time is short, so please check the code and tell me if it is working for you too guys :)
Tomorrow I'm going to try to integrate this code with a GUI.

Thanks!